feat: add ListSaves API and expose it to Lua
By adding this API and exposing it to Lua, we can get rid of the
additional "savefiles" constant within Lua. Instead the Lua can call
this API and rely on the SaveGameManager to correctly access the list
of savegames.

The building-block APIs (ReadDirectory etc) are still extant so a mod can
still override the behaviour if desired.