build: fix invoking scan_enums.py
The "scan_enums.py" script requires a Python v2 interpreter to be
available. The "CMakeLists.txt" simply searched for a Python
interpreter, which failed on modern Debian systems.

Instead of simply searching for "python", the CMake script was adjusted
to specifically search for a Python 2 interpreter, and then to
explicitly invoke that to run the "scan_enums.py" scripts.