feat: fix completion display of scan manager
The scan manager shows the completion percentage of the current scan
target while performing a scan. This is confusing for players when
performing orbital scan missions which require less than 100% of the
planet to be scanned.

For example, the mission might request that 42% of the target planet
needs to be scanned, but the completion percentage always goes from
0..100% - confusing the player when it goes past 42%.

For surface scan missions it was ok since the scan target is in
kilometers, so having a 0..100% completion is understandable.

This change modifies both displays to instead of a completion
percentage, showing the actual value completed so far.

For orbital scans, it is a percentage, which goes from 0..<target>%. For
surface scans, it is in kilometers, going from 0..<target>km.