feat: implemenent onDebugReload for ScanGauge
Other gauges reload as part of the gauges module, but the ScanGuage does
not as it's implemented in a different module.

This commit adds an option 'onDebugReload' member to all guages, and the
gauges module now invokes it if it is implemented.

The ScanGauge is now properly reloaded when the Lua debug reload
keystroke is invoked.