refactor: run SaveGameStats in a job
The idea is to run this expensive task in an asynchronous job
and notify the Lua side once it is completed. This is to prevent
the stutter from occurring every time the SaveGameStats API is
called from Lua.

This commit implements this idea by:

* add a job manager to the LuaManager class which can run any job on
  the Lua job queue.
* add a LoadGameToJsonJob implementation in SaveGameManager, and a new
  function, LoadGameToJsonAsync(), which schedules an instance of this
  job on the Lua job queue.
* call this new function from Lua, as well as registering a callback for
  when the job completes, in saveloadgame.lua.

Anecdotal testing in debug builds shows an impressive speed increase
when loading the list of savegames.