refactor: run SaveGameStats in a job
The idea is to run this expensive task in an asynchronous job
and notify the Lua side once it is completed. This is to prevent
the stutter from occurring every time the SaveGameStats API is
called from Lua.

In practise, the jobs all seem to complete almost instantaneously
which then results in a mass of LuaEvents sent to Lua at the same
time. So instead of a set of small stutters, the interface
experiences one big stutter.

So it appears as if the cause of the stutter wasn't the disk access
and JSON parsing in the C++ side, but the logic invocation and data
transfer between C++ and Lua.

As always, it's better to profile first and optimise later..