feat: change scan mission units to square kilometers
Update the display and the mission generator to generate both surface
and orbit scan missions to use square kilomters.

This improves the display and also simplifies some of the calculations.

While these changes should not have impacted the time required to
complete a mission, additional play-testing is required to compare
the changes against the original implementation and ensure that all
missions remain achievable.