feat: improve return location for scan missions
Set the "desc.returnLocation" field on mission completion to show the
"Set Return Route" button once the mission is finished, and continue
to show the mission description. Also show the return location, if
known, as part of the mission description.

TODO: Further improve the display by breaking out the return location
text based on whether "dropoff" is set.