feat: Scan manager use square kilometres for completion
The Scan manager used by Scout missions shows the completion percentage
of the current scan target while performing a scan. This is potentially
confusing for players when performing orbital scan missions which require
less than 100% of the planet to be scanned.

For example, the mission might request that 42% of the target planet
needs to be scanned, but the completion percentage always goes from
0..100% - confusing the player when it goes past 42%.

Surface scan missions do not have this issue as the scan target is
displayed in kilometres, so having a 0..100% completion is
understandable.

This change modifies both displays to show both the scan target as well
as the completion in square kilometres, eliminating the potentially
confusing percentages. It also adds a progress bar for the currently
active scan. This is to both show a more intuitive way to show to the
player how far along they are in completing the scan target, as well as
to provide faster feedback as the current completion amount can be slow
to update.

While these changes should not have impacted the time required to
complete a mission, additional play-testing is required to compare
the changes against the original implementation and ensure that all
missions remain achievable.