feat: add buy/sell prices to the Commodity Market.
Feature request #5903 raised by lolo101.

The advertised price in the Commodity Market represents the
"mid-market rate", not the actual buy or sell prices. This information
is not very useful to a trader, who has to click on a commodity and then
buy or sell 1 unit to see the actual price.

This change replaces the "mid-market" rate display with the actual buy
and sell prices.