feat: middle-align the column headings in the Commodity Market
A natural improvemnt of the previous commit was to move the alignment
calculation code into a helper function, which can also middle-align
text. This new feature was then used to middle-align the table column
headings.

This function should probably be added to a utility library where it can
be used by other table-based UIs.