feat(radar): auto-zoom to cargo cannisters
Similar to auto-zooming on ships, add the capability for the radar to
auto-zoom on cargo containers. This is very useful when scooping multiple
cargo containers in an area, as otherwise the radar is liable to zoom out
to infinity in-between scooping containers. This makes it very difficult
to lock on to the next container.

TODO: the radar should be switchable between auto and manual modes, with
the manual mode being controllable by the keyboard/joystick. A radar
mode and range indicator might also be useful to add.

In future, different radar equipment could provide different ranges
and/or capabilities.