feat(radar): add icons for the new radar buttons
Szlrd created icons both for the scanner/radar toggle button as well as
for the manual/automatic zoom mode indicator button.

Two of the new icons are on the 21st row (starting on index 320) for the
radar's automatic and manual zoom mode indicator.

This necessitated modifying several other files to reflect the increased
grid size of the icons.svg file.

Co-authored-by: Szlrd <bszlrd@gmail.com>