refactor(views): sanitize Lua view names
Make all string representations of the views follow the same pattern, and
give views which did not have a name a name.

Following this, use that string representation in the Lua side as well
instead of using a separate string. This allows for unique and easy
referencing of views across the entire codebsae using 'grep' or
Visual Studio Code's search function.

This also allowed for some simplification of getting/setting views.