feat(radar): capture mouse wheel for zooming radar
By placing all of the radar drawing routines into a window, the mouse
wheel is captured and is not passed back to the world view.

The original implementation failed to get this to work properly as the
wrong API was used for setting the center of the 3D radar. Using
'setCursorScreenPops()' instead of 'setCursorPos()' has the 3D radar
background showing up in the correct location even when placed inside a
window.