ui(spin): fix deselecting objects in SystemView
With the new spinning code using Ctrl+LMB as an alternative way to spin
the view, a selected object in the SystemView was deselected whenever
the user tries to spin the view.

By also checking against the Ctrl key when selecting/deselecting objects
in the SystemView this issue is resolved.

TODO: unify the modifier key with Input::Manager::IsMouseRotatePressed()