ui(spin): fix WorldView spinning and weapon discharge
Mouse rotation is restricted to only when the relevant bindings are
active; that is, either Ctrl+LMB, or MMB. If another mouse-button is
pressed at the same time, it no longer counts as activating rotation.

It still allows for additional modifiers to be pressed, which is required
when used in conjunction with the rotation speed modifiers.

This prevents the issue from having the ShipView rotating while also
firing the laser.

ALSO: unify rotation speed modifiers (in all but the Model Spinner).
ALSO: use the Left-Alt key instead of the Right-Shift key for the
secondary speed modifier; using both Shift-keys is a bit cumbersome.