Fixing icons.svg
Fixing the icons.svg I provided for the radar PR, I put the new icons for that onto an older version of the file.
The nice thing is that the free space is similar in both, so there's no need to do anything in the theme, if I'm reading it correct.