feat(radar): improve checking when mouse is over radar area
The 2D radar is trivial as it is a circle, but the 3D scanner is an
ellipse. For now we check if the mouse is in a rectangle surrounding
the scanner, which is an improvement but not a complete fix.

TODO: if the mouse leaves the radar area while the popup for selecting
between the radar modes is displayed, the popup disappears. The
mouse-area-check should be disabled while the popup is displayed.