feat(radar): clicking on a radar "pip" sets or clears it as target
Left-clicking on a "pip" sets it as the combat target if it is a ship,
otherwise as the navigation target. If the "pip" is the current target,
the target is cleared instead.

When there are multiple targets represented by a single "pip", a target
selection popup is shown allowing the playing to select the correct
target.

Fix an issue if the player clicks outside a popup which could get the mouse
handling confused and prevent future interactions with the radar.

Also hide the "pip" tooltip if the target selection popup is active.