fix(equipment-outfitter): show additional equipment details
The following additional information is now rendered in the equipment
card:
- equipment stock level
- equipment tech level

It should be trivial to add more equipment details to the display now,
such as weapon damage or firing rates etc.

Additionally, out-of-stock items are now rendered instead of hidden.

This allows players to see all items which are normally available at a
station instead of showing a partial list which could lead to player
confusion when they can't find a particular equipment item even at high
tech-level stations.