feat(equipment-outfitter): render stock level in equipment details
As well as on the equipment item card, render the stock level in the
equipment details tab as well to make it more obvious.

There's potential for rendering more information about the equipment as
well, such as Tech Level, to make it easier for the player to search for
out-of-stock equipment, for example.