Sort lua components during serialization
In the serialization of the Lua objects there is a system of "one
definition", when all subsequent mention of the object only refer to it.
It is vital that the full definition meets before the rest of the refs
to the object, otherwise an error occurs.

The same objects can now meet in different lua components, and this
means that we must ensure the order of processing procedure during
saving and loading.

fixes #6064