Balance Scout mission ads (#6095)
The Scout module creates on average two new ads every day. If the
player sits on or around the same station for a month or more this
will eventually build up to over a hundred ads. This PR slows down
the creation of new ads by one sixth to about one new ad every three
days.

In response to #5760