Character - Separate first name and surname
We need to access the names separately but they were previously gene-
rated together and saved as one in Character.name, with the result of
having to go through the hoops of separating the names which is really
hard if the name is composed of more than two names.

Now they are generated as two separate strings, Character.firstname and
Character.surname. Character.name is still present.