fix(imgui): fix disabling ctrl-tab window switching
The original implementation to disable window switching no longer works
with the new version of ImGui and caused the game to stop with an
assertion failure during startup.

This was broken during PR #6143.

After some research a new approach was discovered, as discussed and
documented in the ImGui issue tracker:
- https://github.com/ocornut/imgui/issues/4828
- https://github.com/ocornut/imgui/issues/3255