fix(imgui): fix renderFaceGenButtons
The new version of imgui no longer allows certain window flags when
creating a child window and instead requires corresponding child window
flags to be specified.

The groundwork was laid in PR #6143, but not all code has been updated.

In the case of the "AlwaysAutoResize" flag, this must now be a child window
flag and it must be used in conjunction with "AutoResizeX" and/or
"AutoResizeY".