fix(imgui): fix debug crash and use symbols for draw flags
Fix crash bug in debug builds when calling "addRect" and "addRectFilled"
with legacy values for the DrawFlags. Instead of exposing the new draw
flags via a symbol table which would incur excessive overheads for these
APIs we continue to use the legacy values in the LUA and convert to the
new values in the C++ interface, as was already implemented for
"addRectFaded".

Also convert all magic numbers used in "addRect", "addRectFilled", and
"addRectFaded" calls referencing the draw flags to use symbols.