fix: select star for hyperspace target
Fixes #6142

When setting the hyperspace target, a check is in place which skips
updating the target if it is in the same system as the current target.

This prevents changing the destination star in a multi-star system for the
next jump. No such check exists when editing a jump further along a route.

The fix is to modify the check and only skip the update if the new
hyperspace target is identical to the current target and not just whether
they are in the same system.