feat(SecondHand): clearer error messages
Fix #6130 by introducing additional error messages which detail why the
equipment can not be installed in the ship. The module now distinguishes
between the ship not having a compatible equipment slot, not having a free
slot, or not having enough free space.

The secondhand module now also disables ads which the player cannot
purchase as well as giving the reason when the player clicks on a disabled
ad. Previously the player had to interact with the seller and click on buy
before the error was shown.