feat(events): add onFocusGained and onFocusLost LUA events
As suggested by sturnclaw, implement checking for the relevant SDL
events in the GuiApplication class, then exposing it via a virtual
function to be overridden in the Pi::App class.

The events are then raised on the UI event queue rather than the main
game event queue.

In order to handle these events in the LUA side, the following code
is required:

```lua
ui.Events.Register("onFocusLost", function() ... end)
```

Co-authored-by: Axtel Sturnclaw <sturnclaw@protonmail.com>