feat(events): add onFocusGained and onFocusLost LUA events
TODO: This is probably not the best place to implement detecting the window
focus lost/gained status and sending out the corresponding LUA event, but
"It Works For Me(tm)".

Request suggestions on where to implement this instead. I have investigated
the "opengl/RendererGL.cpp" which would require also exposing this to the
generic "Renderer" class in order to be accessible from elsewhere.

Implementing it directly in the "RengererOGL" class is infeasible as it
does not know about LUA.