feat(settings): implement generic setters/getters for settings
On the suggestion of sturnclaw, implement generic getters and setters for
game settings.

Future work will rework existing settings where appropriate to use this
approach.

Note that this removes the ability for a settings getter/setter to invoke
additional logic in the C++ side; these are a straight pass-through to the
settings file. So, for example, these can't be used to mute/unmute the
audio as that invokes more functionality.