fix(audio): fix muting/unmuting effects and music
When muting, the audio volume is set to 0 and the current audio volume
is saved to the settings.

When unmuting, the audio volume is set to the current audio volume (ie, 0)
instead of the audio volume saved in the settings, effectively resulting
in the audio still being muted.

Fixed by setting the volume to whatever was last saved into the settings.