feat(ui): target radial support hyperspace clouds
Implement a long-running TODO item in the target radial popup menu.

When the target is a departure hyperspace cloud and the player has a
hyperspace cloud analyzer fitted, add a target option to set the players'
hyperspace target to the destination system of the cloud.

TODO: update the route planner to reflect setting the hyperspace
destination.