fix(ShipMarket): correct parameters for matching slot types
The logic in the ShipMarket display invoked the "SlotTypeMatches" function
with inverted parameters, leading to incorrect results when matching slot
patterns.

For example, searching for "missile_bay" slots fails to return any,
because they are all specialised such as "missile_bay.opli_internal". With
the correct parameter order the correct number is returned.