feat(ui): update ImGuiCol definitions for LUA
Update the LUA ImGui color definition enum table to match the current
ImGui definitions.

Presumably this table was not updated following an upgrade of the
DearImGui library, or alternatively, these missing entries were never
defined in the past as they weren't being used.

In either case, ensure that the LUA side has full access to the color
definitions on the C++ side.