feat(ui-systemview): place game-wide settings into a section
Instead of using a "meta-section" (SystemView.Xxx) to save the system view
game settings, expand on the game configuration getters and setters to
include the section, and then pass "SystemView" as the section for all the
systemview game configuration options.

Also force saving the settings on game end. For some reason, during inital
testing, the settings were always saved on exiting the game, but when
re-testing, they weren't. So now force it.

Really the game process should automatically save any unsaved settings
when it exits.